disabled profile panels for mobile devices
Added Mobile_Detect.php (see link in humans.txt) to determine if user is
on mobile device. If they are, clicking one of the profile links with a
panel takes them straight to the page like a normal link. Panels still
slide out for desktop users.
Also changed wrapper span #main to a div.